backend/requirements.txt
braintree==4.18.1
Django==3.1.7
djangorestframework==3.14.0
